port/engine/enhancements: finish the warning cleanup to zero
Sign-compare casts, UNUSED locals, explicit switch defaults, override
keywords, [[maybe_unused]] fields, lambda capture trims, M_PI guard,
and a dropped no-op statement in MoleGroup (the mirror flip happens in
OMole's own constructor).

+++ src/engine/TrainCrossing.cpp
@@ -52,7 +52,7 @@ void TrainCrossing::AICrossingBehaviour(s32 playerId) {
     if ((!(D_801631E0[playerId] != 0)) ||
         (set_vehicle_render_distance_flags(gPlayers[playerId].pos, TRAIN_CROSSING_AI_DISTANCE, 0))) {
         if ((OnTriggered == 1) && ((Timer) > FRAMES_SINCE_CROSSING_ACTIVATED)) {
-            if ((sSomeNearestPathPoint > WaypointMin) && (sSomeNearestPathPoint < WaypointMax)) {
+            if (((s32) sSomeNearestPathPoint > (s32) WaypointMin) && ((s32) sSomeNearestPathPoint < (s32) WaypointMax)) {
                 bStopAICrossing[playerId] = 1;
             }
         }
